Update the DELETE route of the blogsRouter to deny access to the user who didn't create the blog. Refactored the PUT route

+++ controllers/blogs.js
@@ -36,23 +36,32 @@ blogsRouter.post("/", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token missing or invalid" });
+  }
+  const user = await User.findById(decodedToken.id);
+  const blog = await Blog.findById(request.params.id);
+
+  if (blog.user.toString() !== user._id.toString()) {
+    return response.status(401).json({ error: "access denied" });
+  }
+
+  user.blogs = user.blogs.filter((id) => id.toString() !== request.params.id);
+
   await Blog.findByIdAndRemove(request.params.id);
+  await user.save();
   response.status(204).end();
 });
 
 blogsRouter.put("/:id", async (request, response) => {
-  const body = request.body;
-
-  const blog = {
-    title: body.title,
-    url: body.url,
-    author: body.author,
-    likes: body.likes,
-  };
-
-  const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, {
-    new: true,
-  });
+  const updatedBlog = await Blog.findByIdAndUpdate(
+    request.params.id,
+    request.body,
+    {
+      new: true,
+    }
+  ).populate("user", { username: 1, name: 1 });
   response.json(updatedBlog);
 });
